Rename select.tsx to Select.tsx
